# Compacted time step report