One small step in making output more consistent