@alfbr 2020-09-23
Before:
```
Time step 0, stepsize 3 days, at day 299/316, date = 01-Sep-1998
Time step summary: well its =  1, newton its =  2, linearizations =  3 ( 0.912 sec), linear its =  41 ( 1.285 sec)
```
After:
```
Time step 1, stepsize 3 days, at day 1/8, date = 07-Nov-1997
Well its= 1, Newton its= 3, linearizations= 4 (0.3sec), linear its= 43 (0.3sec)

```

After discussion with some of my colleagues it was clear that this line in its completeness is not understood or read by ordinary users no matter what we do. Personally I really us the inforamtion, both timings of linearizations and timing of linear iterations. Removing the timing of the assembly will make the line significantly less informative to me. One option is to move the whole line to the PRT file. I will miss it, but I realize that that is of minor importance.

With respect to your wish of keeping spaces between "its" and "=", I disagree. But if you really feel strongly about it, I will put them back in. Also, please see what cases you have where setw(2) is insufficient. If it only happens very rarely, I would rather have the line a little longer when it happens.